requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
httpie==2.4.0
idna==3.2
Markdown==3.3.4
Pygments==2.9.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing==3.7.4.3
urllib3==1.26.6
